refactor(connectors): extract response config sanitization
